Fetch spark image from S3 object store
We've precompiled the spark.qemu image and uploaded it to
the Amazon S3 object store where anyone can download it.
With this commit we provide a deployment that uses the spark.qemu
from S3 instead from the imageserver, as is the case in
spark-local.yaml.

Please note that from now on it's possible to deploy our Spark
cluster on Kubernetes WITHOUT building both the OSv (provided on S3)
and the mikelangelo/spark-driver (provided on DockerHub).

Signed-off-by: Miha Pleško <miha.plesko@xlab.si>